+++ src/tests/browser/render-mesh.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2019 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2019-2021 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
  */
@@ -15,6 +15,8 @@ import { Mesh } from '../../mol-geo/geometry/mesh/mesh';
 import { Color } from '../../mol-util/color';
 import { createRenderObject } from '../../mol-gl/render-object';
 import { Representation } from '../../mol-repr/representation';
+import { Torus } from '../../mol-geo/primitive/torus';
+import { ParamDefinition } from '../../mol-util/param-definition';
 
 const parent = document.getElementById('app')!;
 parent.style.width = '100%';
@@ -31,17 +33,24 @@ function meshRepr() {
     const builderState = MeshBuilder.createState();
 
     const t = Mat4.identity();
-    MeshBuilder.addCage(builderState, t, HexagonalPrismCage(), 0.005, 2, 20);
+    Mat4.scaleUniformly(t, t, 10);
+    MeshBuilder.addCage(builderState, t, HexagonalPrismCage(), 0.05, 2, 20);
 
     const t2 = Mat4.identity();
-    Mat4.scaleUniformly(t2, t2, 0.1);
+    Mat4.scaleUniformly(t2, t2, 1);
     MeshBuilder.addPrimitive(builderState, t2, SpikedBall(3));
+
+    const t3 = Mat4.identity();
+    Mat4.scaleUniformly(t3, t3, 8);
+    MeshBuilder.addPrimitive(builderState, t3, Torus({ tubularSegments: 64, radialSegments: 32, tube: 0.1 }));
 
     const mesh = MeshBuilder.getMesh(builderState);
 
-    const values = Mesh.Utils.createValuesSimple(mesh, {}, Color(0xFF0000), 1);
-    const state = Mesh.Utils.createRenderableState({});
+    const props = ParamDefinition.getDefaultValues(Mesh.Utils.Params);
+    const values = Mesh.Utils.createValuesSimple(mesh, props, Color(0xFF4433), 1);
+    const state = Mesh.Utils.createRenderableState(props);
     const renderObject = createRenderObject('mesh', values, state, -1);
+    console.log('mesh', renderObject);
     const repr = Representation.fromRenderObject('mesh', renderObject);
     return repr;
 }