+++ src/tests/browser/render-shape.ts
@@ -111,6 +111,7 @@ const repr = ShapeRepresentation(getShape, Mesh.Utils);
 export async function init() {
     // Create shape from myData and add to canvas3d
     await repr.createOrUpdate({}, myData).run((p: Progress) => console.log(Progress.format(p)));
+    console.log('shape', repr);
     canvas3d.add(repr);
     canvas3d.requestCameraReset();